Require cryptsetup >=2.4
