+++ .github/workflows/run-tests.yml
@@ -12,7 +12,7 @@ jobs:
         with:
           go-version: '1.20'
       - run: dnf install -y gcc pkg-config cryptsetup-libs cryptsetup-devel
-      - run: go test -tags="cryptsetup2.4" -v ./...
+      - run: go test -v ./...
   ubuntu-22-04-go-1-18:
     runs-on: ubuntu-22.04
     steps:
